+++ docs/changelog.md
@@ -1,6 +1,12 @@
 ## unreleased
 
 - `NEW` Allow simulation of extracted FMUs and pre-loaded model descriptions
+- `NEW` Re-load an FMU in the GUI
+- `NEW` Load start values from an FMU
+- `NEW` Write changed start values to FMU
+- `NEW` Table editor for 1-d and 2-d array variables
+- 'FIXED' Logger callback for FMI 1.0
+- `FIXED` Handling of `None` in setString()
 
 ## v0.2.2 (2018-03-13)